feat: add support for riscv64
The macos_compat layer should also include a `JMP_BUF_SIZE` const (or a
totally different structure, just like with glibc), but it has been
omitted because there is no support for RISC V at the moment.

The asm based implementation should be fine for riscv32, but I did not
check what `JmpBufFields` should be.